This PR is mainly about merging the currently separate implementation of TransformerState::select() with the other two select() versions. The difference between Transformer (batch-major) and the other (RNN, hard-att: time-major) is now reflected in the lowest-level function, State::select().

This PR includes several steps:
 * const correctness
 * streamlining the DecoderState::select() functions
 * re-implementing the original TransformerState::select() code inside State::select(), and more efficiently using reshaping

It also includes a handful of fixes:
 * hard-att did not seem to setPosition()
 * a few warnings (size_t -> int).

+++ src/training/graph_group_async.cpp
@@ -41,7 +41,7 @@ void AsyncGraphGroup::fetchParams(Tensor oldParams,
   int pos = 0;
 
   std::vector<std::thread> threads;
-  for(size_t idx = 0; idx < devices_.size(); idx++) {
+  for(int idx = 0; idx < devices_.size(); idx++) {
     threads.emplace_back(std::thread(
         [&](int idx, int pos) {
           // individual mutex per-shard
@@ -64,7 +64,7 @@ void AsyncGraphGroup::pushGradients(Tensor newGrads,
   // add instead of copy?
   std::vector<std::thread> threads;
   int pos = 0;
-  for(size_t idx = 0; idx < devices_.size(); idx++) {
+  for(int idx = 0; idx < devices_.size(); idx++) {
     threads.emplace_back(std::thread(
         [&](int idx, int pos) {
           // individual mutex per-shard
